On 64 bit, subtracting two pointers produces a long result. On 32 bit, it's an int. Even though long _is_ 32 bits on a 32 bit systems, gcc warns about it because reasons.
Also, the warning being that "expects int, but type is wchar_t"... no, type
is not wchar_t. Type is probably long. Specify the ACTUAL TYPE, not the random
typedef alias for it. If the translated type _did_ match, there wouldn't
be a warning! (This is why c89 promoted all arguments to int, precisely
so this wasn't a problem.)

+++ toys/posix/printf.c
@@ -130,7 +130,7 @@ void printf_main(void)
 
           sprintf(to, "*.*L%c", c);
           printf(toybuf, wp[0], wp[1], ld);
-        } else error_exit("bad %%%c@%ld", c, f-*toys.optargs);
+        } else error_exit("bad %%%c@%ld", c, (long)(f-*toys.optargs));
 
         if (end && (errno || *end)) perror_msg("bad %%%c %s", c, aa);
       }
